% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimergm.R
\name{bimergm}
\alias{bimergm}
\title{Bipartite exponential random graph models with nodal random effects}
\usage{
bimergm(formula, iter)
}
\arguments{
\item{formula}{formula; an \code{R} formula object, of the form
<network> ~ <model terms> where <network> is a \code{\link[network]{network}}
object and <model terms> are \code{\link[ergm]{ergm.terms}}. The terms
\code{b1sociality} and \code{b2sociality} are added to the formula argument.}

\item{iter}{count; number of iterations used for iterative estimation
of the model coefficients and the random effects, 10 iterations are usually
enough.}
}
\description{
Function to fit ERGMs with nodal random effects to bipartite networks.
Estimation is carried out combining the stepping algorithm
(see \code{\link[ergm]{control.ergm}}) for the fixed parameters and
pseudo-likelihood (see \code{\link[mgcv]{bam}}) for the nodal random effects.
}
\details{
This implemented algorithm can only handle undirected bipartite
networks.
}
\references{
Sevag Kevork, Göran Kaeurmann (2021).
Bipartite exponential random graph models with nodal random effects
}
\author{
Sevag Kevork
}
